services/blog/requirements.txt
fastapi==0.100.0
uvicorn==0.23.1
sqlalchemy==2.0.19
pydantic==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.11.1
pytest==7.4.0
httpx==0.24.1
python-dotenv==1.0.0
uuid==1.30
python-slugify==8.0.1
